+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Python 3.4 and 3.5 support
 - pypy support
+- Support comparison of OrderedDict. Refs #55
 
 
 ## No previous changelog history.